LayoutTests/fast/viewport/ios/shrink-to-fit-content-constant-width-expected.txt

This test verifies that a page with a constant width viewport smaller than the actual view width is scaled to fit the view. To test manually, load the page and verify that the bar spans the full width of the page.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS minScale is expectedScale
PASS innerWidth is 300
PASS successfullyParsed is true

TEST COMPLETE
